+++ lib/signal_strength_indicator.dart
@@ -1,3 +1,1 @@
-library signal_strength_indicator;
-
 export 'src/signal_strength_indicator.dart';